Major library refactor
- Moved all header files to appropriate subfolders to be context-isolated
- Used `target_include_directories` instead of `include_directories` to better manage which library connects to where, and make this project include-able as a sub-project to bigger projects
- Made all `#include` statements system-level and correctly prefixed
