+++ src/st-trace/trace.c
@@ -10,12 +10,12 @@
 #include <time.h>
 #include <unistd.h>
 
-#include <stlink.h>
+#include <stlink/stlink.h>
 
-#include <chipid.h>
-#include <logging.h>
-#include <read_write.h>
-#include <register.h>
+#include <stlink/chipid.h>
+#include <stlink/logging.h>
+#include <stlink/read_write.h>
+#include <stlink/register.h>
 #include <usb.h>
 
 #define DEFAULT_LOGGING_LEVEL 50
