+++ src/stlink-lib/include/stlink/calculate.h
