+++ src/stlink-lib/include/stlink/chipid.h
