+++ src/stlink-lib/include/stlink/md5.h
@@ -7,7 +7,7 @@
 #ifndef MD5_H
 #define MD5_H
 
-#include "map_file.h"
+#include <stlink/map_file.h>
 
 void md5_calculate(mapped_file_t *);
 void stlink_checksum(mapped_file_t *);
