+++ src/stlink-lib/src/common_flash.c
@@ -10,15 +10,15 @@
 #include <string.h>
 #include <unistd.h>
 
-#include <stlink.h>
-#include "common_flash.h"
+#include <stlink/stlink.h>
+#include <stlink/common_flash.h>
 
-#include "calculate.h"
-#include "flash_loader.h"
-#include "logging.h"
-#include "map_file.h"
-#include "md5.h"
-#include "read_write.h"
+#include <stlink/calculate.h>
+#include <stlink/flash_loader.h>
+#include <stlink/logging.h>
+#include <stlink/map_file.h>
+#include <stlink/md5.h>
+#include <stlink/read_write.h>
 
 #define DEBUG_FLASH 0
