+++ src/stlink-lib/src/read_write.c
@@ -8,10 +8,10 @@
 #include <stdio.h>
 #include <string.h>
 
-#include <stlink.h>
-#include "read_write.h"
+#include <stlink/stlink.h>
+#include <stlink/read_write.h>
 
-#include "logging.h"
+#include <stlink/logging.h>
 
 // Endianness
 // https://commandcenter.blogspot.com/2012/04/byte-order-fallacy.html
