+++ tests/flash.c
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include <stlink.h>
+#include <stlink/stlink.h>
 #include <flash.h>
 #include <flash_opts.h>
